requirements.txt
codecarbon==2.3.4
numpy==1.23.5
Pillow==11.0.0
prefetch_generator==1.0.3
ptflops==0.7.4
Requests==2.32.3
scikit_learn==1.4.2
scikit_learn==1.1.3
torch==2.0.1
torch==1.11.0
torch==2.3.1
torch==2.2.1
torch==2.5.1
torchvision==0.18.1
torchvision==0.17.1
torchvision==0.20.0
